requirements.txt
# Django
Django==5.0.1
django-environ==0.11.2
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.1.0

# Telegram Bot
aiogram==3.3.0
aiohttp==3.9.1

# Database
psycopg2-binary==2.9.9
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# QR Code generation
qrcode[pil]==7.4.2
Pillow==10.2.0

# Utilities
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1

# Image processing
Pillow==10.2.0

# Date handling
python-dateutil==2.8.2

